requirements.txt
aiohttp>=3.8.0
loguru>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
textblob>=0.17.1
vaderSentiment>=3.3.2
websocket-client>=1.6.0
python-binance>=1.0.19
ta>=0.10.0
streamlit>=1.29.0
plotly>=5.18.0
altair>=5.2.0
yfinance>=0.2.33
scipy>=1.11.4
scikit-learn>=1.3.2
python-dateutil>=2.8.2 